feat: add abstractions for drizzle
