+++ packages/plugin-drizzle/src/schema/types.ts
@@ -25,9 +25,3 @@ export const numberTimestamp = customType<{ data: number; driverData: string }>(
         },
     }
 );
-
-export interface IDatabaseClientManager<T> {
-    initialize(): Promise<void>;
-    getConnection(): T;
-    close(): Promise<void>;
-}
